feat: migrate collections to new module, rename PIAF to AF
